+++ Source/SelfService/Web/spaces/applications/applicationsList.tsx
@@ -7,22 +7,7 @@ import { ShortInfoWithEnvironment } from '../../apis/solutions/api';
 
 import { List } from '@mui/material';
 
-import { Button } from '@dolittle/design-system';
-
-// Make buttons as wide as the longest one
-const styles = {
-    list: {
-        p: 0,
-        mt: 2.5,
-        display: 'inline-block',
-    },
-    button: {
-        mb: 2,
-        minWidth: 155,
-        minHeight: 36,
-        display: 'block',
-    },
-};
+import { ApplicationsListItem } from './applicationsListItem';
 
 export type ApplicationsListProps = {
     data: ShortInfoWithEnvironment[];
@@ -30,18 +15,12 @@ export type ApplicationsListProps = {
 };
 
 export const ApplicationsList = ({ data, onChoose }: ApplicationsListProps) =>
-    <List sx={styles.list}>
-        {
-            data.map(application => (
-                <Button
-                    variant='filled'
-                    key={`${application.id}-${application.environment}`}
-                    label={`${application.name}-${application.environment}`}
-                    isFullWidth
-                    onClick={() => onChoose(application)}
-                    sx={styles.button}
-                />
-            ))
-        }
-        <Button label='Create new Application' variant='outlined' isFullWidth />
+    <List>
+        {data.map(application =>
+            <ApplicationsListItem
+                key={`${application.id}-${application.environment}`}
+                application={application}
+                onChoose={() => onChoose(application)}
+            />
+        )}
     </List>;
